migrate to stream sink
